Cleanup: Remove legacy CommonJS test files after full ESM migration and test porting. All chains now use ESM and pass Vitest.
