--- server/agents/chains/design/LoopClarifierChain.js
@@ -1,32 +1,0 @@
-const { LLMChain, PromptTemplate } = require('langchain');
-const fs = require('fs');
-const path = require('path');
-
-function createLoopClarifierChain(llm) {
-  const promptPath = path.join(__dirname, '../../prompts/design/loop-clarifier.md');
-  let promptString;
-  try {
-    promptString = fs.readFileSync(promptPath, 'utf8');
-  } catch (err) {
-    throw new Error(`Prompt file not found: ${promptPath}`);
-  }
-  const prompt = new PromptTemplate({
-    template: promptString,
-    inputVariables: ['title', 'pitch']
-  });
-  const chain = new LLMChain({ llm, prompt });
-
-  return {
-    async invoke(input) {
-      if (!input || typeof input !== 'object' || !input.title || !input.pitch) {
-        throw new Error('Input must be an object with title and pitch');
-      }
-      const result = await chain.call(input);
-      if (!result.loop || typeof result.loop !== 'string') {
-        throw new Error('Output missing required loop string');
-      }
-      return result;
-    }
-  };
-}
-module.exports = { createLoopClarifierChain };
